tests: Port multi_thread_sighandler test to catch2
The old test had quite a lot of infrastructure that is already
implemented by the TestServer class used by the catch2 tests, so the
result is fairly lean.

Note that the test is not fully reproduce the full extent of what we
previously had in the cxxtest version, since it's missing the
registration of signal handlers before the Tango system is initialised
in the server.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>