Utils/ThSig: Move getpid from ThSig to Utils::misc_init
This code was unnecessarily complicated, since a process' PID is the
same for all threads, including the signal thread, where the PID was
being calculated. The PID was then made available to the DServerSignal
object on the main thread, which required synchronisation. Finally,
DServerSignal made this information available, but only Utils::misc_init
ever queried it.

This commit removes all PID-related operations from both ThSig and
DServerSignal, moving the getpid() call directly to the Utils::misc_init
function.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>