DServerSignal: Remove inheritance from TangoMonitor
Now that there is no communication between the DServerSignal and its
inner ThSignal thread, there is no need to inherit from TangoMonitor,
which was used to synchronise several operations between the two.

This commit removes this inheritance. As a result, we don't need to
provide an explicit ctor anymore, and can default it instead.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>