AttributeProxy: Don't delete the device proxy object on connection failures
In 91107f58 (Merge branch 'offer_obs_support_to_idl_4_and_5_devices_#1193'
into 'main', 2023-12-13) we switched from using connect to using reconnect
in a lot of places. This was done to have more metadata from the DS.

A side-effect of that change is that when you try to call routines which
reconnect you can get an Tango::DevFailed exception if you hit the rate
limit, where you formerly only got Tango::ConnectionFailed exceptions.

AttributeProxy's real_constructor function did not take that into account.
It must only delete the pointers, and also null them to ease debugging, if
the exception was not the rate limit one.

The added test triggers a crash without the fix.