Connection::set_timeout_millis: Only set timeout for device
The `device_6`, `device_5` etc. are all related to `device` via calls to
`_narrow()` (~CORBA's `dynamic_cast`) which means they all refer to the
same `CORBA::Object`. In other words, they are pointers, all with the
same value, just with different types.

As `setClientCallTimeout` accepts a `CORBA::Object_ptr` it can only
operate on the base `CORBA::Object`, this means that we do not need to
call `setClientCallTimeout` for the more derived objects as the call
using `device` sets the timeout for `CORBA::Object` itself and all the
pointers we have to the object will use this timeout, regardless of
their type.

The test added here demonstrates that this is the case because
`DeviceProxy::read_attribute` ends up as a call to
`device_5->read_attributes_5()`, `device_4->read_attributes_4()` etc.
depending on the tests `idlver`. If we weren't setting the timeout
correctly for the `device_5` then we would not expect a timeout because
the 500ms sleep is much shorter than the default 3s timeout.