tests/catch2/platform/impl_win32: Do not include SIGINT as a relevant signal
We create the TestServer as a separate process group by passing
CREATE_NEW_PROCESS_GROUP to CreateProcess(), so that we can send a
SIGBREAK to this process by passing its pid to
GenerateConsoleCtrlEvent() in stop_server(). However, the
documentation[1] for GenerateConsoleCtrlEvent() says the following about
passing "CTRL_C_EVENT" (SIGINT) to this function:

> Generates a CTRL+C signal. This signal cannot be limited to a specific
process group. If dwProcessGroupId is nonzero, this function will
succeed, but the CTRL+C signal will not be received by processes within
the specified process group.

If we pass 0 for dwPorcessGroupId it will send the CTRL+C to every
process "sharing the console", however, CREATE_NEW_PROCESS_GROUP means
we don't share a console with the server[2], so we have no way of
sending the SIGINT without changing how to manage the server process.

Give all this, let's just not test the SIGINT behaviour on Windows as
there isn't much difference between SIGINT and SIGBREAK in thsig.cpp
since d41ecdb5 (DServerSignal: Put special value to request signal
thread termination, 2024-12-22).

[1] https://learn.microsoft.com/en-us/windows/console/generateconsolectrlevent
[2] https://learn.microsoft.com/en-us/windows/win32/api/processthreadsapi/nf-processthreadsapi-createprocessw