Util::shutdown_ds: Do not cleanup logging
Tearing down the logging should be the final action that we do on the server, however, shutdown_ds() this is not the final action of the server.

shutdown_ds() is called from either ThSig or KillThread, which wakes up the main thread once the ORB is shutdown.  The main thread then joins the two other threads in server_cleanup, before finally exiting.  We should be tearing down the logging system _after_ these threads have joined because those threads might log something before the exit.  This is only really an issue when using the Kill command to kill the server because we have to signal to ThSig that it needs to shut down and it logs after receiving a new command from the queue.  If ThSig shutsdown the server this isn't an issue, because it just exits straight away after shutdown_ds() is called and there is no KillThread in this scenario.  Rather than requiring that these threads don't log, let's instead just make sure that the logging system is the last part we tear down.

On Windows logging from the ThSig while the logging system is being teared down from a Kill command can result in a deadlock.  Moving the logging teardown to server_cleanup() fixes this.