catch2/platform/impl_unix.cpp: Use ppoll() on Linux
When running all the Catch2 tests in a single process, i.e. Catch2Tests
without any arguments, we exceed FD_SETSIZE (1024) resulting in stack
corruption and a segfault.

This commit switches to using ppoll() instead of pselect() on Linux as
there is no max fd limit for ppoll().  However, ppoll() is only
supported on Linux, so we must still use pselect() on macOS.  This means
the issue FD_SETSIZE still persists on macOS.