catch2/platform/impl_unix: Remove kill() override
We should not exit the test application if sending a single fails,
instead we should raise an exception to be caught and reported by
catch2.

Moreover, we do not want the errors reported by kill() in stop_server()
function as the error checking will be handled later in wait_for_stop().