EventConsumer::unsubscribe_event: Don't unsubscribe during shutdown
Since 36281159 (- Add automatic event un-subscription in DeviceProxy class
dtor, 2016-06-10) we unsubscribe from events in the DeviceProxy dtor.

While this is a good idea in general it can create a deadlock if we are
currently shutting down the event system. This deadlock can happen if the
ZMQ control socket is stopped earlier, via ZmqEventConsumer::cleanup_EventChannel_map()
than the DelayEvent messages from EventConsumer::unsubscribe_event. We
then wait forever in the DelayEvent ctor for the recv message.

Now we could have a timeout for the recv message in the DelayEvent
ctor/dtors. But this could introduce other issues when the control socket
is really busy and the timeout too short.

But if we are about to shutdown the event system, there is actually no
point in unscubscribing so let's just skip it in this case.