catch2_configure_zmq_ports: Allow any ephemeral port
On Windows, there is no upper limit to the value of the ephemeral port
by default, so the previous value of 60999 can fail[1].

The 60999 value comes from the default ephemeral port range on Linux,
however, we don't want the tests to only pass on systems with the
default configuration.

I don't think it is actually worth the effort to check that we got back
a value in the ephermal port range.  Let's trust that the OS does the
correct thing and will report the failure if it cannot find an ephermal
port for us. As such, let's just require that it is a valid 16-bit integer.

[1] e.g. https://gitlab.com/tango-controls/cppTango/-/jobs/9531679883