.gitlab-ci.yml: Skip old tests on most Linux jobs
The only Linux job where we still run the old tests is debian12.  This
means that the coverage job doesn't include the old tests, however, this
is probably an improvement as we want new MRs to add new catch2 tests
and so the coverage only from the catch2 tests is more useful.

We still build the tests on jobs which don't run the tests, to make sure
the tests at least build everywhere.