ci/print_coredumps.sh: Don't require specific core pattern
We were using the core pattern to know what the executable name was for
the particular core file.  However, this is unnecessary because the
executable name is stored in the corefile itself.  gdb can exact this
and is smart enough to go find the executable to look for debug symbols.

We have been doing this dumb core pattern stuff since 2c34c489
(.gitlab-ci.yml: Print backtraces if we find core dumps, 2023-06-15).