catch2_test_dtypes: Test for overflow in rds routine
There is an overflow in the method to check rds, we store the difference
of 2 int type in the same type.
Added tests that perform just this using min and max values of each
types.
There are tests for float and double as well but they shouldn't
overflow. We could fall into issues due to rounding with float
arithmetic though…