w_attribute.cpp: check_rds_alarm code refactor to fix overflow issue
The code in check_rds_alarm was factorized to fix the issue only once.
With this the delta is now stored in a big enough primitve and we do not
have an overflow.
Note that in attribute.h, get_value for Attr_CheckValue was made const,
as in practice it is, let write code more ccnst correct.