server/w_attribute: check_rds fix integer overflow for long and long64
For signed integer, the auto promotion to compute the delta is not
enough for 'big' integer, we need to detect that we are going to have an
overflow, and in this case set the rds to alarm, cause if it overflows,
it's bigger than any limit we could set.

BTW setting a limit on the rds delta would save us a lot of trouble
there, on top of making a lot of sense.