server/device_3_templ.h: move init_* to device_3.cpp
This templated functions are only used in device_3.cpp, use them there in
an anonymous namespace.
Note that init_out_data had to be adapted to use member field dev_attr,
and the opportunity was taken to fix a bit const correctness of
init_polled_out_data.
device_3.h, get rid of the functions.