server/attribute: move some functions to anonymous namespace set_upd_properties, check_range_coherency
these methods were in attrgetsetprop, and attribute_templ, all moved to
anonymous namespace in attribute.cpp, and set/get_properties moved to
attribute.cpp.
Templates for set_upd_properties were removed from attribute.h in favor
of declaration of all accepted overloads.
For get/set_properties, the template is kept, but is_tango_base_type is
used to protect for misuse of the interface.
As a consequence they are not needed in the templ_inst.cpp file either.
Note that we used the opportunity to make the argument of set_properties
const as it should.