w_attribute: move set/get_min/max_value to implementation file
This ended up being more work than planned, since this used functions
that were protected in attribute.h but that we started moving to
anonymous namespace.
In the end we created an internal header attribute_utils.h to hold this
helper functions that are shared, a lot more of these functions were
found, so in the end depleted attribute.h of the following functions:
 - delete_startup_exception
 - throw_err_format
 - throw_incoherent_val_err
 - throw_err_data_type
 - prop_in_list
 - avns_in_db
 - avns_in_att
All these can be now found in the internal/server/attribute_utils.h, and
implemented in the corresponding cpp.
In w_attribute.h the set/get_min/max_value template function was
protected from misuse with is_tango_base_type.
the function prop_in_list was refactored, to be more concise.
The avns_in_* names are bad, but having no clue what avns stands for
they will stay like this, they're not exposed anymore at least.
This is a big commit but by introducing the
internal/server/attribute_utils.h file we can further factorize and
internalize code in attribute.h.