Headers, remove #include <tango/tango.h>
It's a complicated work to break in small steps. The approach here is to
remove a tango/tango.h include and fix everything till it works again.
Due to the lot of cyclic dependencies it touching a lot of files.

In this step the starting point was client/api_util.cpp.

Apart from removing the include and replacing it with smaller ones, the
other big modification was to move string_dup and string_free, before
in tango.h, to tango_const.h, as it is widely used.