server/dserver.h: remove tango/tango.h include
Move TANGO_BASE_CLASS define from tango_const to its own header
tango_base_class.
This generate a lot of cyclic dependencies that are more easily resolved
with putting this in a dedicated header file.
On top of the define, the proper device header is included, so when
writing a device server you can include this header to get the define
and the proper include in the same go.