server/attribute.cpp: protect set_value with is_tango_base_type
This way we prevent anyone from calling set_value with an unsupported
type.
Note that there is an overload for DevEncoded as there is an explicit
specialization of set_value_date_quality that uses it, preventing us
from defining it in the implementation file.

move set_value then from attribute_templ.h to attribute implementation.