ci/windows-build.ps1: Use correct MSVC runtime library
Since a long time in cppTango we use the static MSVC runtime library for
cppTango's static library and the dynamic one for the shared library.

The same approach has to be taken by every of our dependencies, especially if we
allocate heap memory in one library and destroyed it in another one.

Broken since 2057969d (Fix appveyor build on windows without using vcpkg., 2021-09-29).

This is also the reason for the crash we are seeing on windows in the jpeg
tests.

See #1296.