.windows-gitlab.yml: Fix windows build
For compiling libjpeg-turbo we need the Netwide Assembler (NASM).
Unforunately the domain nasm.us which we are gathering it from is down. As
we had problems in the past already, we now use one which is hosted on
known location.

The zip was downloaded from https://github.com/ilammy/setup-nasm/blob/mirror/releasebuilds/2.15.05/win64/nasm-2.15.05-win64.zip.

We also use the x64 variant for both compilation architectures, as the
architecture of the nasm compiler does not have to match the architecture
we are compiling to.