Windows support for the replacement of tango/tango.h
Problems mainly came from the usage of 'struct timeval' in tango_clock.h
because the struct was no longer included. On windows it can be included
either from 'winsock.h' or 'winsock2.h'. I decided to include it from
'winsock2.h' and it caused a lot of issues because sometimes 'winsock.h'
was included just before it through the include of 'windows.h'.

The problem was transparent before when we were using tango.h because it
happened to properly include 'winsock2.h' before any 'windows.h'. Now
that we no longer have a common header to includes everything we are
required to properly include it everywhere it's needed.