tests/catch2_create_cpp_class.cpp: test cpp class loading.
Add a test to check that a call to create_cpp_class can find a library
that has been provided without its full name.

Create a DummyClass library to be loaded. It has 2 targets, one without
prefix, one with prefix.

Tests are disabled on windows cause the dynamic loader search strategy
cannot be altered by an env variable, making this tests quite
complicated to run.

They are disabled on static build as well cause we need to make sure we
actually have this functionnality with static build.