EventConsumerKeepAliveThread: Remove unnecessary sleep
This sleep causes short lived processes which subscribe to events, such
as tests, _very_ slow to exit.

Removing this sleep does not seem to cause any issues detected by our
tests and it is unclear why this is required at all as the very first
thing we do in the loop is a 10 second `timedwait()` on a condition
variable.

Waiting using a condition variable is preferable as this sleep can be
woken during the sleep to stop the thread.