EventConsumerKeepAliveThread::stateless_subscription_failed: Make it work again
Currently we don't send out an event when a stateless subscription failed
during resubscription.
The reason is that we tried to match an unqualified event name like `idl5_change`
with our event names without IDL prefix.

This is now fixed. In addition we also send out the event_name without IDL
prefix.